pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = "WhatsApp Business Bot implementation"
authors = [
    {name = "Datura",email = "daturasoldev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncio (>=3.4.3,<4.0.0)",
    "aiosmtplib (>=4.0.0,<5.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "json-log-formatter (>=1.1.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flask = "^3.1.0"
requests = "^2.32.3"

